+++ src/main/java/lotto/controller/LottoController.java
@@ -1,4 +1,10 @@
-package lotto;
+package lotto.controller;
+
+import lotto.*;
+import lotto.model.Lottos;
+import lotto.model.Store;
+import lotto.view.InputView;
+import lotto.view.ResultView;
 
 public class LottoController {
     private static final Store store = new Store(Config.LOTTO_ONE_GAME_PRICE);
@@ -9,5 +15,5 @@ public class LottoController {
         Lottos lottos = store.giveLotto();
         ResultView.printLottoNumbers(lottos);
     }
-    
+
 }
