+++ src/main/java/lotto/model/Store.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class Store {
