+++ src/test/java/lotto/model/LottoTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Lotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
